requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.9
python-dotenv==0.9.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
